+++ src/types/index.ts
@@ -3,4 +3,5 @@ export * from "./broadcast";
 export * from "./dokument";
 export * from "./error";
 export * from "./logging";
+export * from "./metrics";
 export * from "./roller";