+++ src/types/roller/RolleType.ts
@@ -1,13 +1,23 @@
-export enum RolleType {
+export enum RolleTypeFullName {
+    BIDRAGSPLIKTIG = "BIDRAGSPLIKTIG",
+    BIDRAGSMOTTAKER = "BIDRAGSMOTTAKER",
+    REELMOTTAKER = "REELMOTTAKER",
+    BARN = "BARN",
+    FEILREGISTRERT = "FEILREGISTRERT",
+}
+
+export enum RolleTypeAbbreviation {
     BM = "BM",
     BP = "BP",
     BA = "BA",
     RM = "RM",
     FR = "FR",
+}
 
+export enum RolleTypeDeprecated {
     BIDRAGS_PLIKTIG = "BIDRAGS_PLIKTIG",
     BIDRAGS_MOTTAKER = "BIDRAGS_MOTTAKER",
-    BARN = "BARN",
     REELL_MOTTAKER = "REELL_MOTTAKER",
-    FEILREGISTRERT = "FEILREGISTRERT",
 }
+
+export type RolleType = RolleTypeFullName | RolleTypeAbbreviation | RolleTypeDeprecated;